Split `CIME.macros_writers` from `CIME.macros`.
This commit also contains documentation improvements for `CIME.macros`.
